Modify the JavaScript file
Converted the 'No of Years' into the actual year in the future.

Wrote a function that reads the value of the range slider and displays the value of the slider to the span tag element adjacent to the slider.

Moved the const variables inside the compute function so that once clicked the function activates and returns the calculation